[#271] Getters for secure and insecure port in common base class.
Eliminate code duplication from implementing the getters for each
specific vertx server (to only cover the ephemeral port determination).
